python/examples/voynich/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voynich-agents"
version = "0.1.0"
description = "LoopAgent — evolutionary population loop primitive for apx-agent"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
keywords = ["databricks", "apx-agent", "evolutionary", "agent", "loop"]
dependencies = [
    "apx-agent>=0.16.0",
    "databricks-sdk>=0.74.0",
    "httpx>=0.27.0",
    "mlflow>=2.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
spark = ["pyspark>=3.5.0"]
dev   = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.20.0",     # httpx mocking
    "pyright>=1.1.380",
    "uvicorn[standard]>=0.34.0",
]

[project.scripts]
voynich-seed   = "apx_agent.workflow.cli:seed_population"
voynich-run    = "apx_agent.workflow.cli:run_generation_batch"
voynich-vacuum = "apx_agent.workflow.cli:vacuum_population"
voynich-export = "apx_agent.workflow.cli:export_top_candidates"

[tool.hatch.build.targets.wheel]
packages = ["apx_agent"]

[tool.pytest.ini_options]
testpaths    = ["tests"]
asyncio_mode = "auto"

[tool.pyright]
include              = ["loop_agent"]
pythonVersion        = "3.11"
typeCheckingMode     = "basic"
reportMissingImports = "warning"
